refactored home-page and about-page, deleted home page contradictions, added carousel to client-review-section, added swiperjs and animation-on-scroll to the project
