+++ package.json
@@ -13,13 +13,16 @@
 		"@fortawesome/fontawesome-svg-core": "^6.1.2",
 		"@fortawesome/free-solid-svg-icons": "^6.1.2",
 		"@fortawesome/react-fontawesome": "^0.2.0",
+		"aos": "^3.0.0-beta.6",
 		"bootstrap": "^5.2.0",
 		"next": "^12.2.4",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
-		"sass": "^1.54.4"
+		"sass": "^1.54.4",
+		"swiper": "^8.3.2"
 	},
 	"devDependencies": {
+		"@types/aos": "^3.0.4",
 		"@types/node": "18.7.1",
 		"@types/react": "18.0.17",
 		"@types/react-dom": "18.0.6",
