+++ src/components/Data.tsx
@@ -1,6 +1,53 @@
 
 import React from "react"
 
+export const ClientReviews = [
+    {
+        id: 1,
+        text: "I had been looking to get my shipment and maintenance done but couldn't find the right Construction Company to do so. Rolanick Construction has been exceptional at what they do, from pre-planning, procurement, constructing in itself and maintenance it has been worth it. They understood the job and they delivered! Thank you!",
+        image: "/assets/clientPic.png",
+        names: "cindy clifford",
+        position: "Creative Director Sobaz Oil and gas"
+    },
+    {
+        id: 2,
+        text: "I had been looking to get my shipment and maintenance done but couldn't find the right Construction Company to do so. Rolanick Construction has been exceptional at what they do, from pre-planning, procurement, constructing in itself and maintenance it has been worth it. They understood the job and they delivered! Thank you!",
+        image: "/assets/clientPic.png",
+        names: "jasper clifford",
+        position: "Creative Director Sobaz Oil and gas"
+    },
+    {
+        id: 3,
+        text: "I had been looking to get my shipment and maintenance done but couldn't find the right Construction Company to do so. Rolanick Construction has been exceptional at what they do, from pre-planning, procurement, constructing in itself and maintenance it has been worth it. They understood the job and they delivered! Thank you!",
+        image: "/assets/clientPic.png",
+        names: "efe clifford",
+        position: "Creative Director Sobaz Oil and gas"
+    }
+]
+
+export const Swipers = [
+    {
+        id: 1,
+        image: "/assets/hero_img1.png"
+    },
+    {
+        id: 2,
+        image: "/assets/hero_img1.png"
+    },
+    {
+        id: 2,
+        image: "/assets/hero_img1.png"
+    },
+    {
+        id: 2,
+        image: "/assets/hero_img1.png"
+    },
+    {
+        id: 2,
+        image: "/assets/hero_img1.png"
+    }
+]
+
 //====Products Data====
 export const Research = [
     {
