+++ src/styles/home.scss
@@ -168,64 +168,100 @@
 	}
 
 	//3rd section
+	.home-section-3 {
+		margin-top: 3rem;
 
-    .thirdSection {
-        margin-top: 7.6875rem;
-        display: grid;
-        grid-template-columns: 1fr 1fr;
-        gap: 4rem;
+		.section3-inner {
+			@media screen and (min-width: $md) {
+				display: flex;
+				justify-content: space-between;
 
+				&-left {
+					width: 47%;
+				}
 
-        .text {
-            margin-top: 9.5rem;
-            // &-inner{
-            //     // .h2{
-            //     //     color: $danger;
-            //     // }
-            // }
-        }
+				&-right {
+					width: 47%;
+				}
+			}
 
-        .box {
-            margin-top: 2.25rem;
+			&-left {
+				h2 {
+					color: $secondary;
+					font-size: 3rem;
+					font-weight: 800;
+					line-height: 3.55556rem;
+				}
 
-            &-inner {
-                display: grid;
-                grid-template-columns: 1fr 1fr 1fr;
+				p {
+					&.text-danger {
+						font-weight: 600;
+					}
 
-                .box-inner-in {
-                    border: 1px solid $primary;
-                    text-align: center;
-                }
+					&.section3-heading {
+						color: $secondary;
+					}
 
-                :nth-child(1) {
-                    border-left: none;
-                }
+					&.section3-paragraph1 {
+						margin-block: 1.5rem 0;
+						font-weight: 500;
+						line-height: 1.77778rem;
+					}
+					&.section3-paragraph2 {
+						margin-bottom: 2.5rem;
+						font-weight: 500;
+						line-height: 1.77778rem;
+					}
+				}
 
-                :last-child {
-                    border-right: none;
-                }
-            }
-        }
+				.info {
+					display: flex;
+					text-align: center;
+					margin-bottom: 1.5rem;
+					@media screen and (max-width: $md) {
+						width: fit-content;
+						margin-inline: auto;
+					}
 
-        .img {
-            display: grid;
-            grid-template-columns: 1fr 1fr 1fr;
-            gap: 2rem;
+					&-sub {
+						padding: 2%;
+						border: 1px solid $primary;
 
-            :nth-child(2) {
-                margin-top: 5rem;
-            }
+						&:nth-child(1),
+						&:nth-child(3) {
+							border-left: none;
+							border-right: none;
+						}
 
-            :last-child {
-                margin-top: 2.5rem;
-            }
-        }
+						p {
+							margin: 0;
 
-        .h1 {
-            font-weight: 600;
-        }
-		
-    
+							&:nth-child(1) {
+								font-size: 3rem;
+								font-weight: 600;
+								color: $primary;
+							}
+
+							&:nth-child(2) {
+								font-size: 1rem;
+								font-weight: 500;
+								color: $secondary;
+							}
+						}
+					}
+				}
+			}
+
+			&-right {
+				height: fit-content;
+				margin-block: auto;
+
+				img {
+					width: 100%;
+					height: auto;
+				}
+			}
+		}
 	}
 
 	//4th section
@@ -414,10 +450,22 @@
 
 	//6th section
 	.home-section-6 {
+		width: 80%;
+
+		.swiper-pagination-clickable .swiper-pagination-bullet {
+			background: $primary;
+			width: 1rem;
+			height: 1rem;
+			border-radius: 50%;
+		}
+			
+			
+
 		.section6-inner {
 			text-align: center;
 			margin-inline: auto;
 			margin-top: 5rem;
+			width: 100%;
 			h2 {
 				color: $secondary;
 				font-size: 3rem;
@@ -542,26 +590,7 @@
 			&-btn {
 				text-align: center;
 				margin-top: 3rem;
-				button{
-					color: white;
-				}
-
 			}
 		}
 	}
-	@media screen and (max-width:500px) {
-        .thirdSection {
-
-            display: flex;
-            flex-direction: column;
-
-            .img {
-                display: flex;
-                gap: 1rem;
-                justify-content: center;
-                width: 100%;
-                height: auto;
-            }
-        }
-    }
 }
